requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
python-multipart==0.0.9
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
